Break apart the monster file
